bravo-homes-platform-main/bravo-qualifier/requirements.txt
httpx>=0.27.0
fastapi>=0.115.0
uvicorn>=0.30.0
openai>=1.40.0
python-dotenv>=1.0.0
